requirements.txt
Django == 1.5.1
django-celery == 3.0.17
celery == 3.0.19

-e git+git@github.com:abourget/gevent-socketio.git@aeece7038b0052ddf6b4228857e4d7a67a6242f2#egg=gevent-socketio
#gevent-socketio == 0.3.5-rc2

djangorestframework == 2.3.5
django-nose == 1.1
gunicorn == 0.17.4
django-compressor == 1.3
django-extensions == 1.1.1
django-jsonfield == 0.9.10
ipython == 0.13.2
ipdb == 0.7
tox == 1.4.3

python-memcached == 1.53
